+++ backend/Models/Beneficiario.js
@@ -5,7 +5,7 @@ const { Schema } = mongoose;
 const beneficiarioSchema = new Schema({
   tipoDocumento: {
     type: String,
-    enum: ['Cedula de ciudadania', 'Tarjeta de identidad'],
+    enum: ['C.C', 'T.I'],
     required: true
   },
   identificacion: {
