--- frontend/src/pages/Ayudantes.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Ayudantes = () => {
-  return (
-    <div>Ayudantes</div>
-  )
-}
-
-export default Ayudantes
