--- frontend/src/pages/Beneficiarios.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import BeneficiariosTabla from '../context/BeneficiariosTabla'
-
-const Beneficiarios = () => {
-  return (
-    <div>
-      <BeneficiariosTabla/>
-    </div>
-  )
-}
-
-export default Beneficiarios
