--- frontend/src/pages/Dashboard.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Dashboard = () => {
-  return (
-    <div>Dashboard</div>
-  )
-}
-
-export default Dashboard
