--- frontend/src/pages/Donaciones.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Donaciones = () => {
-  return (
-    <div className=''>
-      Donaciones
-    </div>
-  )
-}
-
-export default Donaciones
