--- frontend/src/pages/Donadores.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Donadores = () => {
-  return (
-    <div>
-      Donadores
-    </div>
-  )
-}
-
-export default Donadores
