--- frontend/src/pages/Proyectos.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import ProyectosTabla from '../context/ProyectosTabla';
-
-const Proyectos = () => {
-  return (
-    <div className='lg: '>
-      <ProyectosTabla />
-    </div>
-  );
-};
-
-export default Proyectos;
