--- frontend/src/pages/index.jsx
@@ -1,6 +1,0 @@
-export { default as Donaciones } from './Donaciones';
-export { default as Dashboard } from './Dashboard';
-export { default as Donadores } from './Donadores';
-export { default as Beneficiarios } from './Beneficiarios';
-export { default as Ayudantes } from './Ayudantes';
-export { default as Proyectos } from './Proyectos';
